Linux workfile support now matches Windows.
git-svn-id: svn://localhost/public/MoSync/trunk@342 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ runtimes/cpp/core/Core.cpp
@@ -780,7 +780,7 @@ public:
 #define dumpJump(a)
 #endif
 #define JMP_GENERIC(address) dumpJump(address); if(uint(address) >= CODE_SEGMENT_SIZE) {\
-	LOG("\nIllegal jump to 0x%04X\n", address); BIG_PHAT_ERROR(ERR_IMEM_OOB); }\
+	LOG("\nIllegal jump to 0x%04X\n", (uint)address); BIG_PHAT_ERROR(ERR_IMEM_OOB); }\
 		ip = (byte*)(mem_cs + (address));
 #else
 #define JMP_GENERIC(address) \
